refactor: test_one is now common to all platforms
